# Annotate Hash#key? and the likes as inline