--- a/inits.c
+++ b/inits.c
@@ -86,6 +86,7 @@ rb_call_builtin_inits(void)
 {
 #define BUILTIN(n) CALL(builtin_##n)
     BUILTIN(gc);
+    BUILTIN(hash);
     BUILTIN(ractor);
     BUILTIN(numeric);
     BUILTIN(io);