## Benchmark

```
$ benchmark-driver benchmark/mjit_hash.yml -v --rbenv 'before --jit;after --jit'
before --jit: ruby 3.1.0dev (2021-05-22T06:36:56Z master fb4195b969) +JIT [x86_64-linux]
after --jit: ruby 3.1.0dev (2021-05-22T07:31:34Z inline-hash-key f4c65bfdd3) +JIT [x86_64-linux]
Calculating -------------------------------------
                         before --jit  after --jit
hash_include?(hash, key)      28.187M      32.555M i/s -     40.000M times in 1.419075s 1.228708s
 hash_member?(hash, key)      27.858M      32.444M i/s -     40.000M times in 1.435856s 1.232886s
hash_has_key?(hash, key)      28.178M      32.563M i/s -     40.000M times in 1.419551s 1.228397s
    hash_key?(hash, key)      26.726M      32.805M i/s -     40.000M times in 1.496653s 1.219334s

Comparison:
             hash_include?(hash, key)
             after --jit:  32554529.5 i/s
            before --jit:  28187373.4 i/s - 1.15x  slower

              hash_member?(hash, key)
             after --jit:  32444198.8 i/s
            before --jit:  27857944.9 i/s - 1.16x  slower

             hash_has_key?(hash, key)
             after --jit:  32562775.1 i/s
            before --jit:  28177917.8 i/s - 1.16x  slower

                 hash_key?(hash, key)
             after --jit:  32804780.3 i/s
            before --jit:  26726307.0 i/s - 1.23x  slower
```

<details>
<summary>VM performance isn't impacted by this.</summary>

```
$ benchmark-driver benchmark/mjit_hash.yml -v --rbenv 'before;after'
before: ruby 3.1.0dev (2021-05-22T06:36:56Z master fb4195b969) [x86_64-linux]
after: ruby 3.1.0dev (2021-05-22T07:31:34Z inline-hash-key f4c65bfdd3) [x86_64-linux]
Calculating -------------------------------------
                             before       after
hash_include?(hash, key)    24.405M     24.516M i/s -     40.000M times in 1.639013s 1.631577s
 hash_member?(hash, key)    23.984M     24.149M i/s -     40.000M times in 1.667805s 1.656405s
hash_has_key?(hash, key)    25.367M     24.303M i/s -     40.000M times in 1.576830s 1.645916s
    hash_key?(hash, key)    24.978M     24.549M i/s -     40.000M times in 1.601419s 1.629409s

Comparison:
             hash_include?(hash, key)
                   after:  24516154.2 i/s
                  before:  24404934.7 i/s - 1.00x  slower

              hash_member?(hash, key)
                   after:  24148684.2 i/s
                  before:  23983612.7 i/s - 1.01x  slower

             hash_has_key?(hash, key)
                  before:  25367357.0 i/s
                   after:  24302582.1 i/s - 1.04x  slower

                 hash_key?(hash, key)
                  before:  24977841.3 i/s
                   after:  24548783.8 i/s - 1.02x  slower
```

</details>