@k0kubun 2021-05-22
I filed this because I saw https://github.com/Shopify/yjit/blob/7eb8c6ba2c828c4d34b903329c2f63de833e7b2a/yjit_iface.c#L134-L138, but it seems like we do need a frame for this. Never mind.